Renaming classes folder to entities
