+++ src/entities/Criteria.java
@@ -1,4 +1,4 @@
-package src.classes;
+package src.entities;
 
 public class Criteria {
     /**
